Rename IdentityAttestation
